aggiorna rischio utenet

+++ app/src/main/java/it/gadg/contagiapp/modelli/User.java
@@ -4,14 +4,14 @@ public class User {
     public String nome;
     public String cognome;
     public String email;
-    public int rischio;
+    public Long rischio;
     public String etichetta;
 
     public User(String nome,String cognome,String email){
         this.nome = nome;
         this.cognome = cognome;
         this.email = email;
-        this.rischio = 0;
+        this.rischio = 0L;
         this.etichetta="super";
